[INTERNAL] lbt/utils: Rename parseJS => parseUtils
