+++ lib/lbt/analyzer/FioriElementsAnalyzer.js
@@ -60,7 +60,7 @@
 
 const ModuleName = require("../utils/ModuleName");
 const SapUiDefine = require("../calls/SapUiDefine");
-const {parseJS, Syntax} = require("../utils/parseJS");
+const {parseJS, Syntax} = require("../utils/parseUtils");
 const {getValue, isMethodCall, isString} = require("../utils/ASTUtils");
 const log = require("@ui5/logger").getLogger("lbt:analyzer:FioriElementAnalyzer");
