+++ lib/lbt/analyzer/JSModuleAnalyzer.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const {Syntax} = require("../utils/parseJS");
+const {Syntax} = require("../utils/parseUtils");
 const escope = require("escope");
 const ModuleName = require("../utils/ModuleName");
 const {Format: ModuleFormat} = require("../resources/ModuleInfo");
