+++ lib/lbt/analyzer/XMLCompositeAnalyzer.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const {Syntax} = require("../utils/parseJS");
+const {Syntax} = require("../utils/parseUtils");
 const SapUiDefine = require("../calls/SapUiDefine");
 const {getValue, isMethodCall, isString} = require("../utils/ASTUtils");
 const ModuleName = require("../utils/ModuleName");
