+++ lib/lbt/analyzer/analyzeLibraryJS.js
@@ -1,5 +1,5 @@
 "use strict";
-const {parseJS, Syntax} = require("../utils/parseJS");
+const {parseJS, Syntax} = require("../utils/parseUtils");
 const {getPropertyKey, isMethodCall, isIdentifier, getStringArray} = require("../utils/ASTUtils");
 const VisitorKeys = require("estraverse").VisitorKeys;
