+++ lib/lbt/bundle/Builder.js
@@ -4,7 +4,7 @@
 
 const terser = require("terser");
 const {pd} = require("pretty-data");
-const {parseJS, Syntax} = require("../utils/parseJS");
+const {parseJS, Syntax} = require("../utils/parseUtils");
 // const MOZ_SourceMap = require("source-map");
 
 const {isMethodCall} = require("../utils/ASTUtils");
