+++ lib/lbt/resources/ResourcePool.js
@@ -4,7 +4,7 @@
 const fs = require("fs");
 const path = require("path");
 */
-const {parseJS} = require("../utils/parseJS");
+const {parseJS} = require("../utils/parseUtils");
 const ComponentAnalyzer = require("../analyzer/ComponentAnalyzer");
 const SmartTemplateAnalyzer = require("../analyzer/SmartTemplateAnalyzer");
 const FioriElementsAnalyzer = require("../analyzer/FioriElementsAnalyzer");
