+++ lib/lbt/utils/ASTUtils.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const {Syntax} = require("../utils/parseJS");
+const {Syntax} = require("../utils/parseUtils");
 
 /**
  * Checks whether the given node is a string literal.
